fix phan

+++ htdocs/expensereport/class/paymentexpensereport.class.php
@@ -189,7 +189,7 @@ class PaymentExpenseReport extends CommonObject
 
 		$totalamount = 0;
 		foreach ($this->amounts as $key => $value) {  // How payment is dispatch
-			$newvalue = price2num($value, 'MT');
+			$newvalue = (float) price2num($value, 'MT');
 			$this->amounts[$key] = $newvalue;
 			$totalamount += $newvalue;
 		}
